removing actor API
